move top level modules to subpackages
