+++ nominal/nominal.py
@@ -7,7 +7,8 @@ from typing import TYPE_CHECKING, Iterable, Mapping, Sequence
 
 import typing_extensions
 
-from nominal import Connection, _config, ts
+from nominal import Connection, ts
+from nominal.config import _config
 from nominal.core import (
     Asset,
     Attachment,
